+++ bm/Libraries/LibrariesViewController.swift
@@ -192,7 +192,7 @@ extension LibrariesViewController: CLLocationManagerDelegate {
     }
     
     func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        print("\(error)")
+        // print("\(error)")
     }
 }